+++ app/src/main/java/xdsopl/robot36/Phasor.java
@@ -9,7 +9,7 @@ package xdsopl.robot36;
 public class Phasor {
 	private final Complex value;
 	private final Complex delta;
-	Phasor(float freq, float rate) {
+	Phasor(double freq, double rate) {
 		value = new Complex(1, 0);
 		double omega = 2 * Math.PI * freq / rate;
 		delta = new Complex((float) Math.cos(omega), (float) Math.sin(omega));
